+++ src/types/index.ts
@@ -25,3 +25,11 @@ export interface BodyTypographyProps
   italic?: boolean
   underline?: boolean
 }
+
+export type ExcludeNull<T, Keys extends keyof T> = {
+  [K in keyof T]: K extends Keys ? Exclude<T[K], null> : T[K]
+}
+
+export type ExcludeNullAndUndefined<T, K extends keyof T> = {
+  [P in K]-?: NonNullable<T[P]>
+} & Omit<T, K>